feat(ui): inline grammar practice + drop redundant stats tabs (#5 #6)
Grammar detail's practice was a button opening a modal (or navigating to
/lesson), and the Statistics tab duplicated the FSRS metrics already shown
in the title bar. Per rc1 feedback: practice should sit inline on the page
and be hidden when the rule has no quiz; the stats tab is redundant (#5).
Same stats-tab removal on the kanji detail page (#6).

- Extract the practice UI from GrammarPracticeModal into GrammarPracticeSession
  (no Modal wrapper, no close buttons).
- Render the session inline: a full-width section under the desktop grid and
  inside the mobile practice tab. use_media_query mounts it in exactly one
  layout (the session is stateful, so dual-mount would create two independent
  quizzes and collide on test-ids).
- Hide the practice tab/section when the rule has no format_map; drop the
  practice button, GrammarPracticeModal, is_practice_open, and the /lesson nav.
- Remove the Statistics tab from grammar and kanji detail. Collapse FsrsMetrics
  to compact-only: drop FsrsMetricsMode, the Expanded arm, ExpandedMetric, the
  next_review_date prop, and the now-dead next_review bindings.
- E2E: rename grammar-practice-modal -> grammar-practice-session, drop the
  close-button steps (inline has no close), rewrite the four grammar practice
  specs to navigate to the detail page and exercise the inline session.
